(bugfix) make sure cleanup always be executed
this bug is discovered in `respec-given-lab` usage.

The bug is, we call test case's `done` after `runExpectations`,
then `runCleanups`. However, when the very last test case's done is called,
`lab` exit program immediately. so the cleanup of the very last test case
may not be executed!

this commit fixed this bug.